Major project refactoring (slightly broken)
Begin undergoing a complete reboot of the API and server-side code.
The client controller has also changed quite a bit. The API has
simplified quite a bit to use the filesystem instead of MongoDB and
only has 3 REST calls. It no longer stores temporary error files,
but instead immediately returns linting errors as a callback.
(Currently the client only spits them out to the console and does
not display them anywhere.) There's still a lot of issues to get it
to a working state.